Correcciones y funcion delete noticia

+++ src/components/pages/Admin.js
@@ -1,10 +1,10 @@
 import React, { useEffect, useState } from 'react'
-import { MdDeleteForever } from 'react-icons/md';
 import axios from 'axios';
 import { swAlert } from '@sweetalert/with-react';
 import { useNavigate } from 'react-router-dom';
 import EditForm from '../links/EditarForm';
 import AddForm from '../links/AgregarNoticia';
+import DeleteButton from '../links/BorrarNoticia';
 
 
 const Admin = () => {
@@ -16,8 +16,6 @@ const Admin = () => {
 
         let token = sessionStorage.getItem('token');
 
-        console.log(token)
-
         if (!token || token === null) {
             navigate('/');
         }
@@ -25,28 +23,32 @@ const Admin = () => {
         const endPoint = "/noticias";
         axios.get(endPoint).then(response => {
             const data = response.data;
-            console.log(data);
             setNoticias(data);
         }).catch(error => {
             swAlert(<h5>Tuvimos un error, vuelve más tarde</h5>)
         })
-    }, [navigate])
+    }, [navigate, noticias])
 
 
-    { /* const handleDelete = (id) => {
-        const endPoint = `/noticias/borrar_noticia/${id}`;
-        axios.delete(endPoint, {
-            headers: {
-                "Authorization": `Bearer ${token}`
+    const handleDelete = () => {
+        if (confirm("¿Seguro que deseas borrar esta noticia?")) {
+            try {
+                const endPoint = `/noticias/borrar_noticia/${id}`;
+                axios.delete(endPoint, {
+                    headers: {
+                        "Authorization": `Bearer ${token}`
+                    }
+                }).then(response => {
+                    swAlert(<h5>Noticia Borrada</h5>)
+                }).catch(error => {
+                    swAlert(<h5>No se pudo borrar la noticia</h5>);
+                });
+                navigate('/admin');
+            } catch (error) {
+                console.error(`No se pudo borrar la noticia: ${error}`)
             }
-        }).then(response => {
-            swAlert(<h5>Noticia Borrada</h5>)
-        }).catch(error => {
-            swAlert(<h5>No se pudo borrar la noticia</h5>);
-        });
-        navigate('/admin');
+        }
     }
-*/}
 
     return (
         <>
@@ -76,8 +78,10 @@ const Admin = () => {
                                             <th>{noticia._id}</th>
                                             <td>{noticia.titulo}</td>
                                             <td>{noticia.autor}</td>
-                                            <td><button type="button" className="btn btn-danger me-3"><MdDeleteForever className='mb-1' /></button>
-                                                <EditForm noticia={noticia} token={tokenSession} /></td>
+                                            <td>
+                                                <EditForm noticia={noticia} token={tokenSession} />
+                                                <DeleteButton noticia={noticia} />
+                                            </td>
                                         </tr>
                                     </>
                                 )
